ml_module/requirements.txt
torch
torchvision
opencv-python
mtcnn
efficientnet_pytorch
pillow
numpy